chore: move IGetDiskResponse from base.ts to /types
